+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Upcoming Changes
 
 *   Add `has?` alias for `has_selector?`.
+*   Delegate `have` to the RSpec collection matcher when passing an Integer.
 
 ## Capybara Test Helpers 1.0.0 (2020-11-17) ##